refactor: refactor the function to get photo names and URLs, and update the related documentation
The getCasePhotos function in the firebase/storage file originally only retrieves photo URLs.
However, a new feature is being implemented in the cms/edit-case page, allowing users to edit a
previously uploaded case. If the user wants to change the photos, it is necessary to compare the
names of the newly saved photos with the previously saved ones. As a result, we need to update the
getCasePhotos function and make corresponding changes to other related files.